Refactor template system to directory, add CheckedTexture template support, split large modules
Split template.rs into template/mod.rs + template/elements.rs. Add
CheckedTexture and DisabledCheckedTexture to button texture application
during template instantiation (fixes MinimalCheckboxTemplate).

Add ButtonText and EditBox FontString template support. Split large
widget method files into focused modules. Remove animation.rs and
methods_text.rs (merged elsewhere). Various cleanup and type fixes.

+++ src/render/shader/primitive.rs
@@ -137,7 +137,7 @@ impl shader::Primitive for WowUiPrimitive {
         bounds: &Rectangle,
         viewport: &Viewport,
     ) {
-        let scale = viewport.scale_factor() as f32;
+        let scale = viewport.scale_factor();
         let physical_bounds = Rectangle::new(
             iced::Point::new(bounds.x * scale, bounds.y * scale),
             iced::Size::new(bounds.width * scale, bounds.height * scale),
